Edit salesforce connection
