+++ front/components/data_source/CreateConnectionOAuthModal.tsx
@@ -2,7 +2,6 @@ import {
   BookOpenIcon,
   Button,
   CloudArrowLeftRightIcon,
-  Input,
   Page,
   Sheet,
   SheetContainer,
@@ -10,61 +9,29 @@ import {
   SheetHeader,
   SheetTitle,
 } from "@dust-tt/sparkle";
-import {
-  isValidSalesforceClientId,
-  isValidSalesforceClientSecret,
-  isValidSalesforceDomain,
-  isValidZendeskSubdomain,
-} from "@dust-tt/types";
 import Link from "next/link";
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 
 import { useTheme } from "@app/components/sparkle/ThemeContext";
 import type { ConnectorProviderConfiguration } from "@app/lib/connector_providers";
 
-type CreateConnectionConfirmationModalProps = {
+type CreateConnectionOAuthModalProps = {
   connectorProviderConfiguration: ConnectorProviderConfiguration;
   isOpen: boolean;
   onClose: () => void;
   onConfirm: (extraConfig: Record<string, string>) => void;
 };
 
-export function CreateConnectionConfirmationModal({
+export function CreateConnectionOAuthModal({
   connectorProviderConfiguration,
   isOpen,
   onClose,
   onConfirm,
-}: CreateConnectionConfirmationModalProps) {
+}: CreateConnectionOAuthModalProps) {
   const { isDark } = useTheme();
   const [isLoading, setIsLoading] = useState(false);
   const [extraConfig, setExtraConfig] = useState<Record<string, string>>({});
-  const [pkceLoadingStatus, setPkceLoadingStatus] = useState<
-    "idle" | "loading" | "error"
-  >("idle");
-
-  const isExtraConfigValid = useCallback(
-    (extraConfig: Record<string, string>) => {
-      if (connectorProviderConfiguration.connectorProvider === "zendesk") {
-        return isValidZendeskSubdomain(extraConfig.zendesk_subdomain);
-      } else if (
-        connectorProviderConfiguration.connectorProvider === "salesforce"
-      ) {
-        return (
-          !!extraConfig.instance_url &&
-          isValidSalesforceDomain(extraConfig.instance_url) &&
-          !!extraConfig.client_id &&
-          isValidSalesforceClientId(extraConfig.client_id) &&
-          !!extraConfig.client_secret &&
-          isValidSalesforceClientSecret(extraConfig.client_secret) &&
-          !!extraConfig.code_verifier &&
-          !!extraConfig.code_challenge
-        );
-      } else {
-        return true;
-      }
-    },
-    [connectorProviderConfiguration.connectorProvider]
-  );
+  const [isExtraConfigValid, setIsExtraConfigValid] = useState(true);
 
   useEffect(() => {
     if (isOpen) {
@@ -73,50 +40,6 @@ export function CreateConnectionConfirmationModal({
       setExtraConfig({});
     }
   }, [isOpen, setIsLoading]);
-
-  useEffect(() => {
-    async function generatePKCE() {
-      if (
-        connectorProviderConfiguration.connectorProvider === "salesforce" &&
-        isValidSalesforceDomain(extraConfig.instance_url) &&
-        !extraConfig.code_verifier &&
-        pkceLoadingStatus === "idle"
-      ) {
-        setPkceLoadingStatus("loading");
-        try {
-          const response = await fetch(
-            `/api/oauth/pkce?domain=${extraConfig.instance_url}`,
-            {
-              method: "GET",
-              headers: {
-                "Content-Type": "application/json",
-              },
-            }
-          );
-          if (!response.ok) {
-            throw new Error("Failed to generate PKCE challenge");
-          }
-          const { code_verifier, code_challenge } = await response.json();
-          setExtraConfig((extraConfig) => ({
-            ...extraConfig,
-            code_verifier,
-            code_challenge,
-          }));
-          setPkceLoadingStatus("idle");
-        } catch (error) {
-          console.error("Error generating PKCE challenge:", error);
-          setPkceLoadingStatus("error");
-        }
-      }
-    }
-
-    void generatePKCE();
-  }, [
-    connectorProviderConfiguration.connectorProvider,
-    extraConfig.instance_url,
-    extraConfig.code_verifier,
-    pkceLoadingStatus,
-  ]);
 
   return (
     <Sheet
@@ -196,65 +119,12 @@ export function CreateConnectionConfirmationModal({
                 </div>
               )}
 
-              {connectorProviderConfiguration.connectorProvider ===
-                "zendesk" && (
-                <Input
-                  label="Zendesk account subdomain"
-                  message="The first part of your Zendesk account URL."
-                  messageStatus="info"
-                  name="subdomain"
-                  value={extraConfig.zendesk_subdomain ?? ""}
-                  placeholder="my-subdomain"
-                  onChange={(e) => {
-                    setExtraConfig({ zendesk_subdomain: e.target.value });
-                  }}
+              {connectorProviderConfiguration.oauthExtraConfigComponent && (
+                <connectorProviderConfiguration.oauthExtraConfigComponent
+                  extraConfig={extraConfig}
+                  setExtraConfig={setExtraConfig}
+                  setIsExtraConfigValid={setIsExtraConfigValid}
                 />
-              )}
-
-              {connectorProviderConfiguration.connectorProvider ===
-                "salesforce" && (
-                <>
-                  <Input
-                    label="Salesforce instance URL"
-                    message="The URL of your Salesforce organization instance."
-                    name="instance_url"
-                    value={extraConfig.instance_url ?? ""}
-                    placeholder="https://my-org.salesforce.com"
-                    onChange={(e) => {
-                      setExtraConfig((prev) => ({
-                        ...prev,
-                        instance_url: e.target.value,
-                      }));
-                    }}
-                  />
-                  <Input
-                    label="Client ID"
-                    message="The client ID from your Salesforce connected app."
-                    name="client_id"
-                    value={extraConfig.client_id ?? ""}
-                    placeholder="3MVG9..."
-                    onChange={(e) => {
-                      setExtraConfig((prev) => ({
-                        ...prev,
-                        client_id: e.target.value,
-                      }));
-                    }}
-                  />
-                  <Input
-                    label="Client Secret"
-                    message="The client secret from your Salesforce connected app."
-                    name="client_secret"
-                    value={extraConfig.client_secret ?? ""}
-                    placeholder="..."
-                    type="password"
-                    onChange={(e) => {
-                      setExtraConfig((prev) => ({
-                        ...prev,
-                        client_secret: e.target.value,
-                      }));
-                    }}
-                  />
-                </>
               )}
 
               <div className="flex justify-center pt-2">
@@ -267,7 +137,7 @@ export function CreateConnectionConfirmationModal({
                       setIsLoading(true);
                       onConfirm(extraConfig);
                     }}
-                    disabled={!isExtraConfigValid(extraConfig) || isLoading}
+                    disabled={!isExtraConfigValid || isLoading}
                     label={
                       isLoading
                         ? "Connecting..."
