+++ front/components/spaces/AddConnectionMenu.tsx
@@ -34,7 +34,7 @@ import {
 import { useRouter } from "next/router";
 import { useCallback, useState } from "react";
 
-import { CreateConnectionConfirmationModal } from "@app/components/data_source/CreateConnectionConfirmationModal";
+import { CreateConnectionOAuthModal } from "@app/components/data_source/CreateConnectionOAuthModal";
 import { CreateOrUpdateConnectionBigQueryModal } from "@app/components/data_source/CreateOrUpdateConnectionBigQueryModal";
 import { CreateOrUpdateConnectionSnowflakeModal } from "@app/components/data_source/CreateOrUpdateConnectionSnowflakeModal";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
@@ -401,7 +401,7 @@ export const AddConnectionMenu = ({
             case "webcrawler":
             case "gong":
               return (
-                <CreateConnectionConfirmationModal
+                <CreateConnectionOAuthModal
                   key={`${c}-${isOpen}`}
                   connectorProviderConfiguration={CONNECTOR_CONFIGURATIONS[c]}
                   isOpen={isOpen}
